break out m2 config into conftest.py
